mysite/requirements.txt
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
environs==11.0.0
fonttools==4.51.0
kiwisolver==1.4.5
marshmallow==3.21.3
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
